fix: dragger should work with form decorator, close: #5334
