+++ components/upload/Dragger.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Upload from './Upload';
+import { UploadProps } from './interface';
+
+export type DraggerProps = UploadProps & { height?: number };
+
+export default class Dragger extends React.Component<DraggerProps, any> {
+  render() {
+    const { props } = this;
+    return <Upload {...props} type="drag" style={{ height: props.height }}/>;
+  }
+}
